backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pandas
pydantic
python-dotenv
python-dateutil
aiofiles
python-jose[cryptography]
passlib[bcrypt]
